backend/requirements.txt
# FastAPI
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.18

# Supabase
supabase>=2.0.0

# Database (PostgreSQL or local SQLite via LOCAL_DEV_MODE)
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
psycopg[binary]>=3.2.0
aiosqlite>=0.20.0
alembic>=1.14.0

# JWT verification
PyJWT>=2.8.0

# Encryption
cryptography>=42.0.0

# Settings
pydantic-settings>=2.7.0

# Image processing
Pillow>=10.0.0

# AI providers
openai>=1.50.0
google-genai>=1.5.0
replicate>=1.0.0

# HTTP client
httpx>=0.28.0
aiofiles>=24.0.0

# Utilities
python-dotenv>=1.0.0
